I made multiple changes to the web GUI and its functionality: 1 - There is now an option to load CLI scans into the GUI for easier viewing of results. This is done by entering a JSON file from the CLI and a zip of the vunerable files, named in the convention: "folderContaingFile=filename.js." Right now the zip must be manually created. 2 - Due to the unconventional name of the files, where is a fileNameToDirectoryLocation folder that takes the filename and returns it in the format "folderContainingFile/filename.js." Due to the need to sync between the JSON results and the uploaded ZIP, if the full file path is know, this will be printed as well.
Major methods added:
1 - $scope.load()
2 - $scope.matchFileNameFromJSONToZip() --called in load to change full path of file to zip-matchable name using regex, and creates a fullpath var.
3 - $scope.parseLoadedResults() --interperates the JSON results into a format intended to emulate results from a scans run from the browser
4 - $scope.handleFileUploadForLoad() --Handles the upload of the JSON file. The zip file is uploaded using the handle already implemented for the manual scan feature.
5 - UI elements added in scan.html, but their functionality should be clear by reading code.

Future tasks involve making the UI cleaner as all was changed now was for functionality purposes, and updating the CLI to automate the creation of the zip file. Restricting uploads to only JSON and ZIP files also needs to be implemented, although the ZIP reader already reads only .js files from previous implementation. Also, a key note is that the uploaded files and the results from the JSON are synced using the $scope.inputFiles variable, as all communication is done using the filename variable. Hence, if the filenames in $scope.results (JSON is imported to this var) match up with the filenames in $scope.inputFiles (the zip is uploaded here), all synching should be automated.

Note: The folder=filename.js convention is used so that if there are multiple files in the base(parent) directory containing the same file name (i.e. config.js), then the files can be differentiated.

The scanning functionality besides the loading interface should be untouched, besides the fact that as no file path is known for the uploaded files (security reasons), the "Known File Path" will fill with just the repeat of the filename. In the future, look to enable/disable this column in the table based on if the filepath is known.